BLAST fasta genome database problem bugfix
Now fasta files with \r or \n instead of just \n should be good to be used as input.
Changed the storage location from magi_repo/outputs to a file in the same folder as either the fasta or compounds input
A new database is made every time when a fasta input is given. This prevents bugs with different genomes with the same filename.
Changed some file paths to be windows-compatible